requirements.txt
flet>=0.24.0
python-dotenv>=1.0.0
supabase>=2.0.0
weasyprint>=60.0
jinja2>=3.1.0
